junos_facts refactor to work with resources modules
*  Refactor junos_facts module to work with
   network resource module.
